warnings fixed

+++ Lab7/Shapes/Sphere.cpp
@@ -22,12 +22,13 @@ HRESULT Sphere::CreateGeometry(ID3D11Device* m_pDevice)
         for (int j = 0; j <= sliceCount; j++)
         {
             float theta = j * thetaStep;
-            Vertex v = { radius * sin(phi) * cos(theta),
-                                radius * cos(phi),
-                                radius * sin(phi) * sin(theta) };
+            Vertex v = { static_cast<float>(radius * sin(phi) * cos(theta)),
+                                static_cast<float>(radius * cos(phi)),
+                                static_cast<float>(radius * sin(phi) * sin(theta)) };
             Vertices.push_back(v);
         }
     }
+   
     Vertices.push_back({ 0, -radius, 0 });
 
     for (int i = 1; i <= sliceCount; i++) {
@@ -50,7 +51,7 @@ HRESULT Sphere::CreateGeometry(ID3D11Device* m_pDevice)
         }
     }
 
-    int southPoleIndex = Vertices.size() - 1;
+    int southPoleIndex = (int)Vertices.size() - 1;
     baseIndex = southPoleIndex - ringVertexCount;
     for (int i = 0; i < sliceCount; i++) {
         Indices.push_back(southPoleIndex);
@@ -61,7 +62,7 @@ HRESULT Sphere::CreateGeometry(ID3D11Device* m_pDevice)
     D3D11_SUBRESOURCE_DATA data;
 
     D3D11_BUFFER_DESC desc = {};
-    desc.ByteWidth = Vertices.size() * sizeof(Vertex);
+    desc.ByteWidth = static_cast<UINT>(Vertices.size() * sizeof(Vertex));
     desc.Usage = D3D11_USAGE_IMMUTABLE;
     desc.BindFlags = D3D11_BIND_VERTEX_BUFFER;
     desc.CPUAccessFlags = 0;
@@ -69,16 +70,16 @@ HRESULT Sphere::CreateGeometry(ID3D11Device* m_pDevice)
     desc.StructureByteStride = 0;
 
     data.pSysMem = Vertices.data();
-    data.SysMemPitch = Vertices.size() * sizeof(Vertex);
+    data.SysMemPitch = static_cast<UINT>(Vertices.size() * sizeof(Vertex));
     data.SysMemSlicePitch = 0;
 
     HRESULT result = m_pDevice->CreateBuffer(&desc, &data, &m_pVertextBuffer);
 
     if (SUCCEEDED(result))
     {
-        numIndeces = Indices.size();
+        numIndeces = static_cast<int>(Indices.size());
         desc = {};
-        desc.ByteWidth = Indices.size() * sizeof(DWORD);
+        desc.ByteWidth = static_cast<UINT>(Indices.size() * sizeof(DWORD));
         desc.Usage = D3D11_USAGE_IMMUTABLE;
         desc.BindFlags = D3D11_BIND_INDEX_BUFFER;
         desc.CPUAccessFlags = 0;
@@ -86,7 +87,7 @@ HRESULT Sphere::CreateGeometry(ID3D11Device* m_pDevice)
         desc.StructureByteStride = 0;
 
         data.pSysMem = Indices.data();
-        data.SysMemPitch = Indices.size() * sizeof(DWORD);
+        data.SysMemPitch = static_cast<UINT>(Indices.size() * sizeof(DWORD));
         data.SysMemSlicePitch = 0;
 
         result = m_pDevice->CreateBuffer(&desc, &data, &m_pIndexBuffer);
@@ -176,7 +177,7 @@ HRESULT Sphere::CreateGeometry(ID3D11Device* m_pDevice)
 
     for (int i = 0; i < maxInstancesNum; i++)
     {
-        translateMatrices.push_back(DirectX::XMMatrixTranslation(rand() % 10, rand() % 10, rand() % 10));
+        translateMatrices.push_back(DirectX::XMMatrixTranslation((float)(rand() % 10), (float)(rand() % 10), (float)(rand() % 10)));
         scaleMatrices.push_back(DirectX::XMMatrixScaling(1.0f, 1.0f, 1.0f));
         rotateMatrices.push_back(DirectX::XMMatrixRotationAxis(DirectX::XMVectorSet(0.0f, 1.0f, 0.0f, 0.0f), 0.0f));
         geomBuffers.push_back(GeomBuffer());
@@ -228,10 +229,10 @@ void Sphere::Draw(const DirectX::XMMATRIX& projMatrix, const DirectX::XMMATRIX& 
         geomBuffers[i].modelMatrix = scaleMatrices[i] * rotateMatrices[i] * translateMatrices[i];
         geomBuffers[i].modelMatrix = DirectX::XMMatrixTranspose(geomBuffers[i].modelMatrix);
 
-        DirectX::BoundingBox box(DirectX::XMFLOAT3(translateMatrices[i].r[3].m128_f32[0],
-            translateMatrices[i].r[3].m128_f32[1],
-            translateMatrices[i].r[3].m128_f32[2]),
-            DirectX::XMFLOAT3(0.1, 0.1, 0.1));
+        DirectX::BoundingBox box(DirectX::XMFLOAT3((float)translateMatrices[i].r[3].m128_f32[0],
+            (float)translateMatrices[i].r[3].m128_f32[1],
+            (float)translateMatrices[i].r[3].m128_f32[2]),
+            DirectX::XMFLOAT3((float)0.1, (float)0.1, (float)0.1));
         if (fr.Contains(box))
         {
             visibleIndBuffer[visibleObjectNum++].idx.x = i;
